Moved WagtaiLogo out of sidebar modules folder (not a sidebar module)
