+++ client/src/sidebar/components/Sidebar/modules/WagtailBranding.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { gettext } from '../../../../utils/gettext';
 import { ModuleDefinition } from '../Sidebar';
-import WagtailLogo from './WagtailLogo';
+import WagtailLogo from '../../WagtailLogo/WagtailLogo';
 
 interface WagtailBrandingProps {
   homeUrl: string;
