+++ client/src/sidebar/components/WagtailLogo/WagtailLogo.tsx
